pkg/proxy: make authorization concurrent
Now that KRP can take multiple config files, it is possible that it will
need to make multiple outbout SubjectAccessReview requests to the
Kubernetes API. In order to keep authorization fast, this commit
modifies the proxy so that all authorizations are concurrent. As soon as
one authorization fails, the context is cancelled and all other
in-flight calls are stopped.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>