reverse proxy: use Rewrite instead of Director
This is to mitigate Golang's behavior with `ReverseProxy.Director`
when, after the outgoing request is mutated, any headers specified by
"Connection" would get rmoved from the mutated request.
`ReverseProxy.Rewrite()` is only called once the headers specified in
"Connection" were already removed from the outgoing-to-be request.

Rename config file name to authz config filename as it is specific to authz

Rename config file name to authz config filename as it is specific to authz