app,rewrite: remove non-resource attrs
Masking the resource attributes value to "just" path, verb and user is
potentially suboptimal, as it ignores other fields that would be of
interest for a SAR request.
This use-case makes sense for regular HTTP request (non kubernetes
resource request), but those kind of requests, don't populate the other
fields in the resource attributes anyway.
Is theoretically a regression, which we should keep in mind.